Add files via upload
With the development of information technology the data’s are stored in the cloud and it need 
to be protect the privacy of data and management of the data at the same time.   
By these demands the reversible data hiding in encrypted images (RDH-EI) attracts more and 
more researcher’s attention.   
Here propose a novel framework for RDH-EI based on RIT (Reverse image Transformation). 
Here the content of the original image can be encrypted with a text file with the help of key.  
Then the transformed image that looks like the target image is used as the encrypted image, and 
sends to the cloud.   
Therefore, the cloud server can embed data into the encrypted image by using any RDH 
methods for plaintext images.   
RDH-EI is a client free scheme and the data embedding scheme is irrelevant with both process 
encryption and decryption.   
In the proposed method Image framing technique is used to make image frames and these 
frames are embedded with a target image and store into the cloud and also improve the quality 
of encrypted image.   
This works improve the storage capacity cloud as well as the security of data.